# #6058: linker/sourcemap optimization