--- a/.travis.yml
+++ b/.travis.yml
@@ -6,6 +6,5 @@ cache:
     - "dev_bundle"
     - ".meteor"
     - ".babel-cache"
-install: ./meteor --get-ready
 script: TEST_PACKAGES_EXCLUDE="less" ./packages/test-in-console/run.sh
 sudo: false