## Summary

1. The linker split the source code by lines repeatedly (at least twice), since splitting the code into lines is a must, we simply split the lines in the constructor and get the lines from the File object.

2. This PR also rewrites some `_.each` loop, since symbols and lines (N) could be huge, iterating the array elements with the native javascript for loop is faster.

3. ES6 spread operator is more expansive if it's translated into es5 by babel.  push.apply reduces a lot of computation costs.

## Fixes

Issue #6058

### Benchmark Result

The benchmark was tested with our real application, the test command is to clean all linker cache and re-run the build:

```
rm -rf .meteor/local/bundler-cache && METEOR_PROFILE=500 ~/work/meteor/meteor run
```

Before
```
(#1) Total: 18,569 ms (ProjectContext prepareProjectForBuild)
(#2) Total: 175,310 ms (Build App)
```

After
```
(#1) Total: 17,786 ms (ProjectContext prepareProjectForBuild)
(#2) Total: 164,798 ms (Build App)
```

https://gist.github.com/c9s/cc5891c0d9050c7b9a8f47ee7c0f072e