@c9s 2016-07-21
I am not sure why the tests are failing. If I forked from `master`, these tests will pass.