@c9s 2016-07-21
yes, this PR is for optimizing the Meteor tool. I will give you the profiling result tomorrow.