@abernix 2016-07-21
Please open an issue discussing what you plan to fix and showing your profiling then link this PR to it