@c9s 2016-07-22
@abernix This issue is related to https://github.com/meteor/meteor/issues/6058

Rebuild -> Sourcemap Performance