@c9s 2016-07-22
@abernix I've updated the description to include the benchmark result, please take a look.