@c9s 2016-07-22
@abernix do I have to regenerate the dev bundle file everytime I switch to another branch? or I just run ./meteor --get-ready ?