@laosb 2016-07-22
@c9s In your case you don't really need to generate the devbundle since you didn't modify the things in bundle, I think?