@c9s 2016-07-23
Is there a reason that meteor tied to node 0.10 ? I found `--prof-process` is not supported in node 0.10 and it's hard to check the profile result.