@c9s 2016-07-23
The building speed is also improved on travis.

It used to take 20 minutes to finish a build, now it only takes 10 minutes!