@benjamn 2016-07-23
Thanks for working on this!

Can you try pushing more commits at once? Every time you push, it triggers a separate run of the test suite. I'm also happy to cancel the older test runs, but I don't want to step on your toes.