@benjamn 2016-07-23
This pull request really needs to be split up to address individual, specific performance optimizations. #6058 is an umbrella performance issue, not really appropriate as a justification for a pull request.

Please take a step back, think about the different problems you're trying to solve, and open some new issues describing those problems. Then we can have a discussion about which ones should be addressed first, and you can solve each one individually, without making lots of loosely-related changes as you go.

I suspect some of these changes are much more important than others, and I would like to understand why. With everything in one big pull request like this, it's hard to learn any lessons that we can apply when writing code in the future.

I don't want to discourage this kind of work, but it has to be more focused than this. If you're not willing to organize your changes by specific goals, I have to warn you we're not likely to merge them.

Thanks for understanding.