@c9s 2016-07-24
@benjamn Thanks, the reason I pushed the commit atomically is that I broke something in specific commit, so I break the changes into several commits to CI to tests.

The reason I'm pushing the branch with changes together is, it runs really really slow to build the application on our side, therefore I maintained a branch for the team.

I will separate the optimization target in the issue, and rebase the changes onto different branch.